pyproject.toml
[project]
dynamic = ["version"]
name = "edge-proxy"
license = { file = "LICENSE" }
dependencies = [
    "fastapi",
    "flagsmith-flag-engine>=10,<11",
    "flagsmith>=5",
    "httpx",
    "marshmallow",
    "orjson",
    "pydantic",
    "python-decouple",
    "python-dotenv",
    "structlog",
    "uvicorn",
    "pydantic-settings>=2.2.1",
]
requires-python = ">= 3.12"

[project.scripts]
edge-proxy-serve = 'edge_proxy.main:serve'
edge-proxy-render-config = 'edge_proxy.main:render_config'

[tool.rye]
managed = true
dev-dependencies = [
  "pre-commit",
  "pytest",
  "pytest-asyncio",
  "pytest-mock",
  "reorder-python-imports",
  "certifi",
  "pytest-freezegun",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.pytest.ini_options]
addopts = ['-vvvv']
